requirements.txt
Flask
gunicorn
Jinja2
protlib
pytz
requests
urllib3
Werkzeug
requests-oauthlib
mysql-connector-python
configparser